Shifting layout a little
Did a few clean up things and then decided to shift the layout a little
Front "Workout" page will have the list of workouts that you have made should probably be made sortable) if you click on a workout you will be brought to a page for that workout that mimics what this page used to look like (with the graph)

This is a little outside what I was supposed to do, but I want to run with it, I will be sure to leave a page that fits that incase you guys want to scrap this part. Will also be sure to get in contact with Crystal becasue this overlaps with her part a lot.