Done for the flight
So I have a rough start on my idea, and what I should be doing, to make further progress I need internet connection so this will be it for now.

I hope to communicate with the group once I land to get more detials on what I should be doing, otherwise I am going to make links that work, and that might end up being it.